refactor(agentos): sub-daemon cluster co-location ai/daemons/{bridge,kb-alerting,kb-gc,kb-reconciliation}/ + daemon.mjs (#11845)
Sub 9 of Epic #11831 — co-locates the four sub-daemon clusters into per-folder
mirrors of Sub 8's orchestrator pattern.

Source moves (8):
- ai/scripts/bridge-daemon.mjs            → ai/daemons/bridge/daemon.mjs
- ai/scripts/bridge-daemon-queries.mjs    → ai/daemons/bridge/queries.mjs (renamed)
- ai/daemons/KbAlertingService.mjs        → ai/daemons/kb-alerting/KbAlertingService.mjs
- ai/scripts/kb-alerting-daemon.mjs       → ai/daemons/kb-alerting/daemon.mjs
- ai/daemons/KbGarbageCollectionService.mjs → ai/daemons/kb-gc/KbGarbageCollectionService.mjs
- ai/scripts/kb-gc-daemon.mjs             → ai/daemons/kb-gc/daemon.mjs
- ai/daemons/KbReconciliationService.mjs  → ai/daemons/kb-reconciliation/KbReconciliationService.mjs
- ai/scripts/kb-reconciliation-daemon.mjs → ai/daemons/kb-reconciliation/daemon.mjs

Test moves (5 — ticket undercounted; included pre-existing bridge specs):
- bridge-daemon.spec.mjs         → ai/daemons/bridge/daemon.spec.mjs
- bridge-daemon-queries.spec.mjs → ai/daemons/bridge/queries.spec.mjs
- KbAlertingService.spec.mjs          → ai/daemons/kb-alerting/
- KbGarbageCollectionService.spec.mjs → ai/daemons/kb-gc/
- KbReconciliationService.spec.mjs    → ai/daemons/kb-reconciliation/

package.json: 3 npm script paths updated (ai:kb-alerting/kb-gc/kb-reconciliation).

Orchestrator wiring:
- Orchestrator.mjs:14-16  bridge-daemon-queries import → '../bridge/queries.mjs'
- TaskDefinitions.mjs:70  bridgeDaemon.args path → resolve('../daemons/bridge/daemon.mjs')
- TaskDefinitions.mjs:72  expectedCommand → 'daemons/bridge/daemon.mjs' (unique
  among the 5 daemon.mjs files post-Sub-8/9)
- bridge/daemon.mjs:179   enforceSingleton() process.command substring check
  updated to 'daemons/bridge/daemon.mjs' (was 'bridge-daemon.mjs')
- SummarizationCoordinatorService.mjs:8  bridge-daemon-queries import path
- KbAlertRuleEngine + KbGarbageCollectionEngine + KbReconciliationEngine: JSDoc @see paths
- CoalescingEngineService: code reference path
- multi-tenant integration spec:114 dynamic import path

V-B-A discipline from Sub 8 (#11844): swept BOTH static `from '...'` AND
dynamic `await import('...')` call-sites. Local full unit run: 2011 passed,
19 failed (all pre-existing — Authorization, checkAllAgentIdle, checkSunsetted,
QueryReRanker, TextEmbeddingService.retry, WriteSideInvariant, Grid family;
none reference moved files; confirmed via grep).
