+++ ai/daemons/kb-alerting/daemon.mjs
@@ -1,20 +1,20 @@
 /**
- * @module ai/scripts/kb-alerting-daemon
+ * @module ai/daemons/kb-alerting/daemon
  * @summary Thin Node-process boot wrapper for the Phase 4D KB operator-alerting daemon (#11642).
  *
  * Owns Neo namespace bootstrap + SIGTERM / SIGINT clean-stop signal handling +
  * `KbAlertingService.start()` invocation. The class definition (poll loop, rule
- * evaluation, channel dispatch) lives in `ai/daemons/KbAlertingService.mjs`.
+ * evaluation, channel dispatch) lives in `ai/daemons/kb-alerting/KbAlertingService.mjs`.
  *
  * This split matches the canonical Orchestrator class+wrapper pattern: class files in
  * `ai/daemons/` never import Neo; entry-point scripts in `ai/scripts/` own the Neo +
  * core/_export + InstanceManager bootstrap chain.
  *
  * Persistent-process invocation: launchd / systemd should target this script
- * (`node ai/scripts/kb-alerting-daemon.mjs`), or `npm run ai:kb-alerting`. The daemon
+ * (`node ai/daemons/kb-alerting/daemon.mjs`), or `npm run ai:kb-alerting`. The daemon
  * is opt-in — it exits early unless `aiConfig.knowledgeBase.alertingEnabled` is true.
  *
- * @see ai/daemons/KbAlertingService.mjs
+ * @see ai/daemons/kb-alerting/KbAlertingService.mjs
  * @see ai/scripts/orchestrator-daemon.mjs (sibling entry-point wrapper precedent)
  * @see #11642 (Phase 4D operator alerting), #11628 (Phase 4 epic)
  */
@@ -22,12 +22,12 @@
 // Neo namespace bootstrap (entry-point invariant): `Neo` + `core/_export` populate
 // `globalThis.Neo` so any module using `Neo.gatekeep()` / `Neo.setupClass()` at
 // module-load succeeds. `InstanceManager` binds `Neo.find` / `Neo.get` aliases.
-import Neo             from '../../src/Neo.mjs';
-import * as core       from '../../src/core/_export.mjs';
-import InstanceManager from '../../src/manager/Instance.mjs';
+import Neo             from '../../../src/Neo.mjs';
+import * as core       from '../../../src/core/_export.mjs';
+import InstanceManager from '../../../src/manager/Instance.mjs';
 
-import logger            from '../mcp/server/knowledge-base/logger.mjs';
-import KbAlertingService from '../daemons/KbAlertingService.mjs';
+import logger            from '../../mcp/server/knowledge-base/logger.mjs';
+import KbAlertingService from './KbAlertingService.mjs';
 
 const cleanShutdown = signal => {
     logger.info(`[KbAlertingService] Received ${signal}; stopping.`);
