+++ ai/daemons/kb-reconciliation/daemon.mjs
@@ -1,35 +1,35 @@
 /**
- * @module ai/scripts/kb-reconciliation-daemon
+ * @module ai/daemons/kb-reconciliation/daemon
  * @summary Thin Node-process boot wrapper for the Phase 4B KB reconciliation daemon (#11640).
  *
  * Owns Neo namespace bootstrap + SIGTERM / SIGINT clean-stop signal handling +
  * `KbReconciliationService.start()` invocation. The class definition (poll loop, tenant
  * enumeration, config-version diff, telemetry, opt-in tombstone) lives in
- * `ai/daemons/KbReconciliationService.mjs`.
+ * `ai/daemons/kb-reconciliation/KbReconciliationService.mjs`.
  *
  * This split matches the canonical Orchestrator class+wrapper pattern: class files in
  * `ai/daemons/` never import Neo; entry-point scripts in `ai/scripts/` own the Neo +
  * core/_export + InstanceManager bootstrap chain.
  *
  * Persistent-process invocation: launchd / systemd should target this script
- * (`node ai/scripts/kb-reconciliation-daemon.mjs`), or `npm run ai:kb-reconciliation`. The
+ * (`node ai/daemons/kb-reconciliation/daemon.mjs`), or `npm run ai:kb-reconciliation`. The
  * daemon is opt-in — it exits early unless `aiConfig.knowledgeBase.reconciliationEnabled` is
  * true; the destructive auto-tombstone is a second opt-in (`reconciliationAutoTombstone`).
  *
- * @see ai/daemons/KbReconciliationService.mjs
- * @see ai/scripts/kb-alerting-daemon.mjs (sibling wrapper precedent)
+ * @see ai/daemons/kb-reconciliation/KbReconciliationService.mjs
+ * @see ai/daemons/kb-alerting/daemon.mjs (sibling wrapper precedent)
  * @see #11640 (Phase 4B reconciliation daemon), #11628 (Phase 4 epic)
  */
 
 // Neo namespace bootstrap (entry-point invariant): `Neo` + `core/_export` populate
 // `globalThis.Neo` so any module using `Neo.gatekeep()` / `Neo.setupClass()` at
 // module-load succeeds. `InstanceManager` binds `Neo.find` / `Neo.get` aliases.
-import Neo             from '../../src/Neo.mjs';
-import * as core       from '../../src/core/_export.mjs';
-import InstanceManager from '../../src/manager/Instance.mjs';
+import Neo             from '../../../src/Neo.mjs';
+import * as core       from '../../../src/core/_export.mjs';
+import InstanceManager from '../../../src/manager/Instance.mjs';
 
-import logger                  from '../mcp/server/knowledge-base/logger.mjs';
-import KbReconciliationService from '../daemons/KbReconciliationService.mjs';
+import logger                  from '../../mcp/server/knowledge-base/logger.mjs';
+import KbReconciliationService from './KbReconciliationService.mjs';
 
 const cleanShutdown = signal => {
     logger.info(`[KbReconciliationService] Received ${signal}; stopping.`);
