+++ package.json
@@ -36,9 +36,9 @@
         "ai:prune-worktrees"           : "node ./ai/scripts/bootstrapWorktree.mjs --prune-stale",
         "ai:lint-agents"               : "node ./ai/scripts/lint-agents.mjs",
         "ai:lint-skill-manifest"       : "node ./ai/scripts/lint-skill-manifest.mjs",
-        "ai:kb-alerting"               : "node ./ai/scripts/kb-alerting-daemon.mjs",
-        "ai:kb-reconciliation"         : "node ./ai/scripts/kb-reconciliation-daemon.mjs",
-        "ai:kb-gc"                     : "node ./ai/scripts/kb-gc-daemon.mjs",
+        "ai:kb-alerting"               : "node ./ai/daemons/kb-alerting/daemon.mjs",
+        "ai:kb-reconciliation"         : "node ./ai/daemons/kb-reconciliation/daemon.mjs",
+        "ai:kb-gc"                     : "node ./ai/daemons/kb-gc/daemon.mjs",
         "ai:orchestrator"              : "node ./ai/daemons/orchestrator/daemon.mjs",
         "ai:revalidation-sweep"        : "node ./ai/scripts/revalidationSweep.mjs",
         "ai:run-sandman"               : "node ./buildScripts/ai/runSandman.mjs",
